refactor!: Re-write public API to take a `CryptoRng`
This changes the public API to take a `CryptoRng + RngCore` instead of the included AES Rng used for the KATs.
A lot of dependencies were moved to `dev-dependencies`, cleaning up the required deps.

BREAKING CHANGE: This changes the public API.
